pyproject.toml
[project]
name = "dc-partybot"
version = "0.1.0"
description = "一個功能豐富的 Discord 機器人，專為派對、娛樂和社群互動而設計，支援音樂播放、表情符號推薦等功能"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "jimmy071919", email = ""}
]
keywords = ["discord", "bot", "music", "party", "entertainment"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications :: Chat",
    "Topic :: Multimedia :: Sound/Audio :: Players",
]
dependencies = [
    "aiohttp>=3.12.15",
    "discord-py>=2.6.3",
    "google-api-python-client>=2.183.0",
    "google-auth>=2.40.3",
    "google-auth-httplib2>=0.2.0",
    "mutagen>=1.47.0",
    "pycryptodomex>=3.23.0",
    "pynacl>=1.6.0",
    "python-dotenv>=1.1.1",
    "redis>=6.4.0",
    "spotipy>=2.25.1",
    "yt-dlp>=2025.9.23",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "flake8>=7.3.0",
    "mypy>=1.18.2",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
]

[project.scripts]
bot = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
exclude = [
    "*.log",
    "logs/",
    "temp_audio/",
    "*.cookies",
    ".env*",
    "requirements.txt",
    "docker-compose.yml",
    "Dockerfile*",
]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = "-v --tb=short"

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
exclude = [
    ".git",
    "__pycache__",
    "build",
    "dist",
    ".eggs",
    "*.egg-info",
    ".venv",
]
